Move content from schook/userinit to userinit
